Move SimpleRegistry to utils package
